third_party/blink/web_tests/flag-specific/highdpi/accessibility/inline-text-bounds-for-range-br-expected.txt

This is a testharness.js-based test.
FAIL Check bounds of inline text boxes after line breaks assert_approx_equals: Line 1 left expected 8 +/- 1 but got 12
Harness: the test ran to completion.
